+++ app/src/main/java/kr/co/lion/unipiece/ui/buy/BuyFragment.kt
@@ -1,15 +1,19 @@
 package kr.co.lion.unipiece.ui.buy
 
+import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.activity.OnBackPressedCallback
+import androidx.core.view.GravityCompat
 import androidx.viewpager2.widget.ViewPager2
 import com.google.android.material.tabs.TabLayoutMediator
 import kr.co.lion.unipiece.R
 import kr.co.lion.unipiece.databinding.FragmentBuyBinding
+import kr.co.lion.unipiece.databinding.HeaderBuyDrawerBinding
 import kr.co.lion.unipiece.ui.MainActivity
 import kr.co.lion.unipiece.ui.search.SearchFragment
 import kr.co.lion.unipiece.ui.buy.adapter.BuyAdapter
@@ -20,16 +24,12 @@ class BuyFragment : Fragment() {
 
     lateinit var binding: FragmentBuyBinding
 
-    lateinit var mainActivity: MainActivity
-
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
 
         binding = FragmentBuyBinding.inflate(inflater, container, false)
-
-        mainActivity = activity as MainActivity
 
         return binding.root
     }
@@ -38,6 +38,7 @@ class BuyFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
         settingToolbarBuy()
         initViewPager()
+        setBuyNaviDrawer()
     }
 
     fun settingToolbarBuy(){
@@ -47,7 +48,7 @@ class BuyFragment : Fragment() {
 
                 setNavigationIcon(R.drawable.menu_icon)
                 setNavigationOnClickListener {
-                    mainActivity.binding.drawerBuyLayout.open()
+                    binding.drawerBuyLayout.open()
                 }
 
 
@@ -76,6 +77,94 @@ class BuyFragment : Fragment() {
         }
     }
 
+    fun setBuyNaviDrawer(){
+
+        with(binding){
+            if(drawerBuyLayout.isDrawerOpen(GravityCompat.START)){
+                drawerBuyLayout.close()
+            }
+
+            with(navigationDrawer){
+                val headerDrawerBinding = HeaderBuyDrawerBinding.inflate(layoutInflater)
+                addHeaderView(headerDrawerBinding.root)
+
+                headerDrawerBinding.backBtn.setOnClickListener {
+                    drawerBuyLayout.close()
+                }
+
+                setNavigationItemSelectedListener {
+
+                    drawerBuyLayout.close()
+                    it.isChecked = true
+
+                    when(it.itemId){
+                        R.id.menuAll -> {
+
+                        }
+                        R.id.menuArtAll -> {
+
+                        }
+                        R.id.menuArtWest -> {
+
+                        }
+                        R.id.menuArtOri -> {
+
+                        }
+                        R.id.menuArtCalli -> {
+
+                        }
+                        R.id.menuArtSculp -> {
+
+                        }
+                        R.id.menuArtPrint -> {
+
+                        }
+                        R.id.menuArtWood -> {
+
+                        }
+                        R.id.menuArtGlass -> {
+
+                        }
+                        R.id.menuArtFabric -> {
+
+                        }
+                        R.id.menuArtMetal -> {
+
+                        }
+                        R.id.menuArtComic -> {
+
+                        }
+                        R.id.menuArtAni -> {
+
+                        }
+                        R.id.menuHumAll -> {
+
+                        }
+                        R.id.menuHumFiction -> {
+
+                        }
+                        R.id.menuHumPoem -> {
+
+                        }
+                        R.id.menuHumScript -> {
+
+                        }
+                        R.id.menuEngAll -> {
+
+                        }
+                        R.id.menuEngSoft -> {
+
+                        }
+                        R.id.menuEngHard -> {
+
+                        }
+                    }
+
+                    true
+                }
+            }
+        }
+    }
     fun initViewPager() {
 
         val titles = listOf("인기 순", "신규 순")
